Switch from uint64_t to int
While it is prettier in some sense to explicitly state that we intend it to only be nonnegative,
it only provides more issues than anything else. So, let's just completely stick with using int
rather than uint64_t

+++ src/adiar/package_mgr.h
@@ -33,7 +33,7 @@ public:
   inline adiar::bdd leaf_false()
   { return adiar::bdd_false(); }
 
-  inline adiar::bdd exists(const adiar::bdd &b, size_t label)
+  inline adiar::bdd exists(const adiar::bdd &b, int label)
   { return adiar::bdd_exists(b,label); }
 
   inline uint64_t nodecount(const adiar::bdd &b)
